+++ BUILD
@@ -1394,9 +1394,11 @@ grpc_cc_library(
         "//src/core:lib/surface/server.h",
         "//src/core:lib/surface/validate_metadata.h",
         "//src/core:lib/transport/connectivity_state.h",
+        "//src/core:lib/transport/custom_metadata.h",
         "//src/core:lib/transport/error_utils.h",
         "//src/core:lib/transport/metadata_batch.h",
         "//src/core:lib/transport/parsed_metadata.h",
+        "//src/core:lib/transport/simple_slice_based_metadata.h",
         "//src/core:lib/transport/status_conversion.h",
         "//src/core:lib/transport/timeout_encoding.h",
         "//src/core:lib/transport/transport.h",
@@ -1445,7 +1447,6 @@ grpc_cc_library(
         "config",
         "config_vars",
         "cpp_impl_of",
-        "custom_metadata",
         "debug_location",
         "exec_ctx",
         "gpr",
@@ -1536,13 +1537,6 @@ grpc_cc_library(
         "//src/core:useful",
         "//src/core:windows_event_engine",
         "//src/core:windows_event_engine_listener",
-    ],
-)
-
-grpc_cc_library(
-    name = "custom_metadata",
-    hdrs = [
-        "//src/core:lib/transport/custom_metadata.h",
     ],
 )
